\name{courseraPML-package}
\alias{courseraPML-package}
\alias{courseraPML}
\docType{package}
\title{
\packageTitle{courseraPML}
}
\description{
\packageDescription{courseraPML}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{courseraPML}
\packageIndices{courseraPML}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{courseraPML}

Maintainer: \packageMaintainer{courseraPML}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
